gateway-routers: Move to separate crate
