+++ Cargo.toml
@@ -40,6 +40,7 @@ members = [
   "pallets/block-rewards",
   "pallets/connectors",
   "pallets/connectors-gateway",
+  "pallets/connectors-gateway/gateway-routers",
   "pallets/claims",
   "pallets/collator-allowlist",
   "pallets/crowdloan-claim",
