+++ runtime/development/Cargo.toml
@@ -115,6 +115,7 @@ pallet-claims = { path = "../../pallets/claims", default-features = false }
 pallet-collator-allowlist = { path = "../../pallets/collator-allowlist", default-features = false }
 pallet-connectors = { path = "../../pallets/connectors", default-features = false }
 pallet-connectors-gateway = { path = "../../pallets/connectors-gateway", default-features = false }
+gateway-routers = { path = "../../pallets/connectors-gateway/gateway-routers", default-features = false }
 pallet-crowdloan-claim = { path = "../../pallets/crowdloan-claim", default-features = false }
 pallet-crowdloan-reward = { path = "../../pallets/crowdloan-reward", default-features = false }
 pallet-fees = { path = "../../pallets/fees", default-features = false }
